+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"github.com/bleasey/bdns/src/blockchain"
+	"github.com/bleasey/bdns/internal/blockchain"
 )
 
 func main() {
